+++ src/Exceptionless.Web/ClientApp/src/routes/(app)/issues/+page.svelte
@@ -9,8 +9,15 @@
     import * as Sheet from '$comp/ui/sheet';
     import { getEventsByStackIdQuery } from '$features/events/api.svelte';
     import EventsDrawer from '$features/events/components/EventsDrawer.svelte';
+    import { shouldRefreshPersistentEventChanged } from '$features/events/components/filters';
     import EventsDataTable from '$features/events/components/table/EventsDataTable.svelte';
+    import { getTableContext } from '$features/events/components/table/options.svelte';
+    import { type WebSocketMessageValue } from '$features/websockets/models';
     import { persisted } from '$shared/persisted.svelte';
+    import { type FetchClientResponse, useFetchClient } from '@exceptionless/fetchclient';
+    import { createTable } from '@tanstack/svelte-table';
+    import { useEventListener } from 'runed';
+    import { debounce } from 'throttle-debounce';
     import IconOpenInNew from '~icons/mdi/open-in-new';
 
     let selectedStackId = $state<string>();
@@ -53,13 +60,52 @@
     function onFilterRemoved(filter?: IFilter): void {
         persistedFilters.value = filterRemoved(persistedFilters.value, defaultFilters, filter);
     }
+
+    const context = getTableContext<EventSummaryModel<SummaryTemplateKeys>>({ limit: limit.value, mode: 'stack_frequent' });
+    const table = createTable(context.options);
+
+    const client = useFetchClient();
+    let response: FetchClientResponse<EventSummaryModel<SummaryTemplateKeys>[]>;
+
+    async function loadData() {
+        if (client.loading) {
+            return;
+        }
+
+        response = await client.getJSON<EventSummaryModel<SummaryTemplateKeys>[]>('events', {
+            params: {
+                ...context.parameters,
+                filter: ['(type:404 OR type:error)', filter].filter(Boolean).join(' '),
+                time
+            }
+        });
+
+        if (response.ok) {
+            context.data = response.data || [];
+            context.meta = response.meta;
+            table.resetRowSelection();
+        }
+    }
+    const debouncedLoadData = debounce(10000, loadData);
+    async function onPersistentEvent(message: WebSocketMessageValue<'PersistentEventChanged'>) {
+        if (shouldRefreshPersistentEventChanged(persistedFilters.value, filter, message.organization_id, message.project_id, message.stack_id, message.id)) {
+            await debouncedLoadData();
+        }
+    }
+
+    useEventListener(document, 'refresh', async () => await loadData());
+    useEventListener(document, 'PersistentEventChanged', async (event) => await onPersistentEvent((event as CustomEvent).detail));
+
+    $effect(() => {
+        loadData();
+    });
 </script>
 
 <div class="flex flex-col space-y-4">
     <Card.Root>
         <Card.Title class="p-6 pb-0 text-2xl" level={2}>Issues</Card.Title>
         <Card.Content>
-            <EventsDataTable bind:limit={limit.value} {filter} filters={persistedFilters.value} mode="stack_frequent" pageFilter="(type:404 OR type:error)" {rowclick} {time}>
+            <EventsDataTable bind:limit={limit.value} {rowclick} {table}>
                 {#snippet toolbarChildren()}
                     <FacetedFilter.Root changed={onFilterChanged} {facets} remove={onFilterRemoved}></FacetedFilter.Root>
                 {/snippet}
